Create separate page for branching and reverted FormPage.tsx
